refactor(bin-designer): split CutoutWorkspace into focused sibling files
Decompose the 774-LOC CutoutWorkspace.tsx into a slim orchestrator
plus three focused siblings. No behavior change.

- useCutoutWorkspaceCamera.ts (~145 LOC): zoom/pan/container-size state,
  ResizeObserver wiring, fitToView, zoomIn/Out, wheel-zoom-toward-cursor.
- useCutoutWorkspacePointer.ts (~260 LOC): background/move/up handlers,
  marquee selection, middle-click + Space pan, cursor world position,
  Space/Ctrl+0 keyboard shortcuts.
- cutoutWorkspaceContextActions.ts (~175 LOC): pure builder for the
  right-click menu (copy/paste/duplicate/flip/center/lock/etc).
- CutoutWorkspace.tsx (~420 LOC): orchestrator wiring the above together
  with useCutoutInteraction and the WebGL canvas.
